Validate hibernate with db

+++ Backend/src/main/java/swp391/group6/model/Order.java
@@ -2,6 +2,7 @@ package swp391.group6.model;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.List;
 
@@ -25,10 +26,10 @@ public class Order {
     private String shippingAddress;
 
     @Column(nullable = false)
-    private Double shippingFee;
+    private BigDecimal shippingFee;
 
     @Column(nullable = false)
-    private Double discount;
+    private BigDecimal discount;
 
     @Column(nullable = false)
     private Timestamp createdAt;
@@ -67,19 +68,19 @@ public class Order {
         this.shippingAddress = shippingAddress;
     }
 
-    public Double getShippingFee() {
+    public BigDecimal getShippingFee() {
         return shippingFee;
     }
 
-    public void setShippingFee(Double shippingFee) {
+    public void setShippingFee(BigDecimal shippingFee) {
         this.shippingFee = shippingFee;
     }
 
-    public Double getDiscount() {
+    public BigDecimal getDiscount() {
         return discount;
     }
 
-    public void setDiscount(Double discount) {
+    public void setDiscount(BigDecimal discount) {
         this.discount = discount;
     }
